SCMOD-7133: mocks and utils for testing scripts moved to the testing module
