+++ worker-workflow-testing/src/main/java/com/github/cafdataprocessing/workflow/testing/models/NewFailure.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.cafdataprocessing.workflow.models;
+package com.github.cafdataprocessing.workflow.testing.models;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
