+++ worker-workflow-testing/src/main/java/com/github/cafdataprocessing/workflow/testing/models/TaskMock.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.cafdataprocessing.workflow.models;
+package com.github.cafdataprocessing.workflow.testing.models;
 
 import com.hpe.caf.api.worker.WorkerTaskData;
 import com.hpe.caf.worker.document.model.Application;
